[#49] refactor : feature 별로 Policy 인터페이스를 갖도록 변경
- Policy 인터페이스를 common에 위치하면, 추후 Policy의 변경에 따라 모든 feature의 Policy 인터페이스들이 되어야 함에 따라, 각각의 feature 별로 Policy 인터페이스를 갖도록 변경
