+++ src/main/java/com/aorri2/goodsforyou/product/domain/Policy.java
@@ -1,0 +1,4 @@
+package com.aorri2.goodsforyou.product.domain;
+
+public interface Policy<T> {
+}
