+++ src/main/java/com/aorri2/goodsforyou/product/domain/ProductPolicy.java
@@ -1,7 +1,5 @@
 package com.aorri2.goodsforyou.product.domain;
 
-import com.aorri2.goodsforyou.common.policy.Policy;
-
 public interface ProductPolicy extends Policy<Product> {
 }
