+++ src/main/java/com/aorri2/goodsforyou/user/domain/LoginUserPolicy.java
@@ -1,6 +1,5 @@
 package com.aorri2.goodsforyou.user.domain;
 
-import com.aorri2.goodsforyou.common.policy.Policy;
 import com.aorri2.goodsforyou.user.application.command.LoginUserCommand;
 
 public interface LoginUserPolicy extends Policy<LoginUserCommand> {
