+++ src/main/java/com/aorri2/goodsforyou/user/domain/Policy.java
@@ -1,4 +1,4 @@
-package com.aorri2.goodsforyou.common.policy;
+package com.aorri2.goodsforyou.user.domain;
 
 public interface Policy<T> {
 	void apply(T t);
